requirements.txt
requests
pytest
voluptuous
pytest_voluptuous
faker
python-dotenv==0.20.0